+++ Covid19.IndividualService.Tests/Fakes/NoOpLogger.cs
@@ -1,0 +1,11 @@
+using System;
+using Microsoft.Extensions.Logging;
+
+namespace Covid19.IndividualService.Tests.Fakes;
+
+internal class NoOpLogger<T> : ILogger<T>
+{
+    public IDisposable BeginScope<TState>(TState state) => null!;
+    public bool IsEnabled(LogLevel logLevel) => false;
+    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter) { }
+}
