+++ Covid19.IndividualService.Tests/Tests/BookCovidTestCommandHandlerTests.cs
@@ -6,66 +6,15 @@ using AutoMapper;
 using Microsoft.Extensions.Logging;
 using MediatR;
 using MassTransit;
+using System;
+using System.Threading;
+using Covid19.IndividualService.Tests.Fakes;
 using Xunit;
 
 namespace Covid19.IndividualService.Tests
 {
     public class BookCovidTestCommandHandlerTests
     {
-        private class FakeBookingAvailabilityRepository : IBookingAvailabilityRepository
-        {
-            private readonly BookingAvailability? _availability;
-            public FakeBookingAvailabilityRepository(BookingAvailability? availability) => _availability = availability;
-            public Task<BookingAvailability?> FindWhere(System.Linq.Expressions.Expression<Func<BookingAvailability, bool>> filterExpression)
-            {
-                var func = filterExpression.Compile();
-                return Task.FromResult(_availability != null && func(_availability) ? _availability : null);
-            }
-            public Task UpdateAsync(BookingAvailability entity, bool saveCurrentChanges = true) { return Task.CompletedTask; }
-            public Task<BookingAvailability?> GetByIdAsync(Guid id) => throw new NotImplementedException();
-            public Task<BookingAvailability?> GetByIdAsync(int id) => throw new NotImplementedException();
-            public Task<IEnumerable<BookingAvailability>> ListAllAsync() => throw new NotImplementedException();
-            public Task<BookingAvailability> AddAsync(BookingAvailability entity, bool saveCurrentChanges = true) => Task.FromResult(entity);
-            public Task DeleteAsync(BookingAvailability entity, bool saveCurrentChanges = true) => throw new NotImplementedException();
-            public Task<bool> AnyAsync(System.Linq.Expressions.Expression<Func<BookingAvailability, bool>> filterExpression) => throw new NotImplementedException();
-            public Task<IEnumerable<BookingAvailability>> GetPagedResponseAsync(int page, int size) => throw new NotImplementedException();
-            public Task<IEnumerable<BookingAvailability>> WhereAsync(System.Linq.Expressions.Expression<Func<BookingAvailability, bool>> filterExpression) => throw new NotImplementedException();
-            public Task<IEnumerable<TProjected>> ListAllAsync<TProjected>(System.Linq.Expressions.Expression<Func<BookingAvailability, TProjected>> projectionExpression) => throw new NotImplementedException();
-            public Task SaveChangesAsync(bool saveCurrentChanges = true) => Task.CompletedTask;
-        }
-        private class FakeIndividualRepository : IIndividualRepository
-        {
-            public Task<Individual> AddAsync(Individual entity, bool saveCurrentChanges = true) => Task.FromResult(entity);
-            public Task<Individual?> FindWhere(System.Linq.Expressions.Expression<Func<Individual, bool>> filterExpression) => Task.FromResult<Individual?>(null);
-            public Task UpdateAsync(Individual entity, bool saveCurrentChanges = true) => Task.CompletedTask;
-            public Task<Individual?> GetByIdAsync(Guid id) => throw new NotImplementedException();
-            public Task<Individual?> GetByIdAsync(int id) => throw new NotImplementedException();
-            public Task<IEnumerable<Individual>> ListAllAsync() => throw new NotImplementedException();
-            public Task DeleteAsync(Individual entity, bool saveCurrentChanges = true) => throw new NotImplementedException();
-            public Task<bool> AnyAsync(System.Linq.Expressions.Expression<Func<Individual, bool>> filterExpression) => Task.FromResult(false);
-            public Task<IEnumerable<Individual>> GetPagedResponseAsync(int page, int size) => throw new NotImplementedException();
-            public Task<IEnumerable<Individual>> WhereAsync(System.Linq.Expressions.Expression<Func<Individual, bool>> filterExpression) => throw new NotImplementedException();
-            public Task<IEnumerable<TProjected>> ListAllAsync<TProjected>(System.Linq.Expressions.Expression<Func<Individual, TProjected>> projectionExpression) => throw new NotImplementedException();
-            public Task SaveChangesAsync(bool saveCurrentChanges = true) => Task.CompletedTask;
-        }
-        private class NoOpPublishEndpoint : IPublishEndpoint
-        {
-            public List<object> Published { get; } = new();
-            public Task Publish<T>(T message, CancellationToken cancellationToken = default) where T : class
-            {
-                Published.Add(message!);
-                return Task.CompletedTask;
-            }
-            // other methods not used
-            public Task Publish(object message, CancellationToken cancellationToken = default) => Task.CompletedTask;
-            public Task Publish(object message, Type messageType, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        }
-        private class NoOpLogger<T> : ILogger<T>
-        {
-            public IDisposable BeginScope<TState>(TState state) => null!;
-            public bool IsEnabled(LogLevel logLevel) => false;
-            public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter) { }
-        }
         private static IMapper CreateMapper()
         {
             var config = new MapperConfiguration(cfg => cfg.AddProfile<Covid19.IndividualService.Core.Profiles.MappingProfiles>());
